backend/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
branca==0.3.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.2.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
entrypoints==0.3
filelock==3.0.12
folium==0.10.1
idna==2.8
importlib-metadata==1.7.0
importlib-resources==3.0.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.4.2
itypes==1.2.0
jedi==0.17.2
Jinja2==2.10.3
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
nltk==3.5
notebook==6.1.4
numpy==1.17.3
oauthlib==3.1.0
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
pilkit==2.0
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.5
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
pyzmq==19.0.2
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scikit-surprise==1.1.1
scipy==1.5.2
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==20.0.31
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
